shops/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2==2.8.4
gunicorn==20.1.0
